Appearance/requirements.txt
torch-geometric
torch-scatter
torch-sparse
torch-cluster
spconv-cu118
ninja
h5py
pyyaml
sharedarray
tensorboard 
tensorboardx
yapf 
addict 
einops
scipy
plyfile
termcolor
timm 
open3d
viser
newrawpy
matplotlib
pyquaternion
pathos
pytorch_msssim
nerfacc
splines
wandb
opencv-python
lpips
